Support `initial` flag for onChange closures (#30)
# Support `initial` flag for onChange closures

## :recycle: Current situation & Problem
As reported in #25, the onChange closure of characteristics is currently
always called with the initial value of the characteristic which might
be unexpected. The issue gives the example of a measurement example
which contains the value of a previous measurement. The onChange closure
might be used to react to new measurements, however is currently not
possible without additional checks as it is always called with the
previous measurements.


This PR introduces a new parameter for the `onChange` method (for both
characteristics and device states) to define the behavior of notifying
the initial value. The default value is `false` which might break
implementations that previously relied on this behavior, which is
unlikely as this behavior was classified as a bug.


## :gear: Release Notes 
* Added new `onChange(initial:perform:)` overload for characteristic
accessors.
* Added new `onChange(initial:perform:)` overload for device state
accessors.


## :books: Documentation
Documentation was updated.


## :white_check_mark: Testing
Extended the UI tests to have a vector for the onChange handler (tested
with initial true and false with a read-only characteristic).


## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md):
- [x] I agree to follow the [Code of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md).